Moved domain out of invoker
