+++ CMakeLists.txt
@@ -36,8 +36,9 @@ endif ()
 
 # directories
 include_directories (invoker)
-include_directories (utility)
 include_directories (datamanager)
+include_directories (domain)
+include_directories (utility)
 
 if (BUILD_UNITTEST)
 include_directories (mock)
@@ -46,6 +47,7 @@ include_directories (external/gtest/googlemock/include)
 endif()
 
 add_subdirectory (invoker)
+add_subdirectory (domain)
 add_subdirectory (utility)
