+++ application/screen/login/authview.cpp
@@ -22,7 +22,7 @@
 #include <iostream>
 #include <viewcommon.hpp>
 #include <backoffice/dashboard.hpp>
-#include <domain/userlogin/authcontroller.hpp>
+#include <userlogin/authcontroller.hpp>
 #include <logindata/authdata.hpp>
 
 namespace view {
