+++ domain/CMakeLists.txt
@@ -1,0 +1,8 @@
+project (domain)
+
+# Domains
+add_subdirectory (userlogin)
+
+if (BUILD_UNITTEST)
+    add_subdirectory (unittest)
+endif()
