+++ domain/userlogin/interface/authdataif.hpp
@@ -18,8 +18,8 @@
 *           Ben Ziv <pointonsoftware@gmail.com>                                                   *
 *                                                                                                 *
 **************************************************************************************************/
-#ifndef INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
-#define INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
+#ifndef DOMAIN_USERLOGIN_INTERFACE_AUTHDATAIF_HPP_
+#define DOMAIN_USERLOGIN_INTERFACE_AUTHDATAIF_HPP_
 #include <string>
 #include <entity/user.hpp>
 
@@ -28,32 +28,22 @@ namespace authentication {
 /*!
  * Note: If you add/update a function in this interface, please also update the mock class
 */
-class AuthViewIface {
+class AuthDataProviderIface {
  public:
-    AuthViewIface() = default;
-    virtual ~AuthViewIface() = default;
+    AuthDataProviderIface() = default;
+    virtual ~AuthDataProviderIface() = default;
 
-    /**
-    * Will be called if user was found
+    /*!
+     * entity::User findUserByPin(const std::string& pin)
+     * Looks for the user that's assigned with the pin argument
+     * Will return user->pin = entity::User::DEFAULT_PIN if user is not found.
+     *
+     * [in] input pin
+     * [return] user class
     */
-    virtual void loginSuccessful(const entity::User& userInfo) = 0;
-
-    /**
-    * invalid PIN!
-    */
-    virtual void showInvalidPINScreen() = 0;
-
-    /**
-    * user not found
-    */
-    virtual void showUserNotFoundScreen() = 0;
-
-    /**
-    * will be called if something went wrong with the database
-    */
-    virtual void showDataNotReadyScreen() = 0;
+    virtual entity::User findUserByPin(const std::string& inputPin) = 0;
 };
 
 }  // namespace authentication
 }  // namespace domain
-#endif  // INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
+#endif  // DOMAIN_USERLOGIN_INTERFACE_AUTHDATAIF_HPP_
