+++ invoker/CMakeLists.txt
@@ -1,8 +1,4 @@
 project (invoker)
 
-add_subdirectory (domain)
-add_subdirectory (entity)
 
-if (BUILD_UNITTEST)
-    add_subdirectory (unittest)
-endif()
+add_subdirectory (entity)
