--- invoker/domain/CMakeLists.txt
@@ -1,4 +1,0 @@
-project (domain)
-
-# Domains
-add_subdirectory (userlogin)
